pyproject.toml
[tool.poetry]
name = "task-indev"
version = "0.1.0"
description = "A simplified crypto staking platform"
authors = ["Ihor Prokopenko <i.prokopenko.dev@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
drf-yasg = "^1.21.7"
